refactor(Table-component): Refactor onChange event handler
Previously, onChange would return an array on selecting all and an object on selecting an individual

row. This was confusing. Now onChange will return an array for both events.

Any components that use the Table and rely on the onChange prop will need to be updated to handle this change.